Rename rule_set to rules
